backend/requirements.txt
flask==2.3.2
python-docx==0.8.11
pypdf2==3.0.1
openai==0.27.8  # Se for usar OpenAI API
python-dotenv==1.0.0